Create module MoveFilesToDocuments for TryMoveSavedDataLocationIfNeeded
