+++ CMakeLists.txt
@@ -170,6 +170,7 @@ if(WITH_GUI)
   add_subdirectory(src/CodeViewerDemo)
   add_subdirectory(src/ConfigWidgets)
   add_subdirectory(src/DataViews)
+  add_subdirectory(src/MoveFilesToDocuments)
   add_subdirectory(src/Orbit)
   add_subdirectory(src/OrbitGgp)
   add_subdirectory(src/OrbitGl)
