+++ src/MoveFilesToDocuments/include/MoveFilesToDocuments/MoveFilesToDocuments.h
@@ -1,0 +1,14 @@
+// Copyright (c) 2021 The Orbit Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef MOVE_FILES_TO_DOCUMENTS_MOVE_FILES_TO_DOCUMENTS_H_
+#define MOVE_FILES_TO_DOCUMENTS_MOVE_FILES_TO_DOCUMENTS_H_
+
+namespace orbit_move_files_to_documents {
+
+void TryMoveSavedDataLocationIfNeeded();
+
+}  // namespace orbit_move_files_to_documents
+
+#endif  // MOVE_FILES_TO_DOCUMENTS_MOVE_FILES_TO_DOCUMENTS_H_
